Move mention model to reside with message model
