+++ fbchat/models/_message.py
@@ -4,6 +4,29 @@ from __future__ import unicode_literals
 from string import Formatter
 
 
+class Mention(object):
+    #: The thread ID the mention is pointing at
+    thread_id = None
+    #: The character where the mention starts
+    offset = None
+    #: The length of the mention
+    length = None
+
+    def __init__(self, thread_id, offset=0, length=10):
+        """Represents a @mention"""
+        self.thread_id = thread_id
+        self.offset = offset
+        self.length = length
+
+    def __repr__(self):
+        return self.__unicode__()
+
+    def __unicode__(self):
+        return "<Mention {}: offset={} length={}>".format(
+            self.thread_id, self.offset, self.length
+        )
+
+
 class Message(object):
     #: The actual message
     text = None
